Uses shared_ptr to OptimizationProblem inside IpoptSolver
Because we might need to modify the optimization problem (e.g. change
the state bounds) without rebuilding the whole Ipopt object,
we need to use a pointer to an OptimizationProblem in the constructor.
We used to use a reference, but it is simply copied (new
OptimizationProblem(pb)) which means that modifications to the original
OptimizationProblem will not appear in the OptimizationProblem
stored in the IpoptSolver object.

git-svn-id: http://130.66.124.6/svn/eonav/oad_and_monitoring_modules@426 9358c2ab-9c44-400d-ad1c-7153579a8fc0