Can solve azimuth allocation pb
Corrected error in lambda definition (in functors_for_optimizer): the
factor has to be caught by value & not by reference. The reason we were
originally capturing it by reference was that we wanted the factor to be
updated inside the lambda when it was changed outside the lambda (eg.
when multiplying by a constant when calculating a derivative). The
problem is that several temporary objects are created & the address of
the captured factor can become invalid. The solution is to use Node's
update_lambda method to redefine the lambda each time we're changing the
multiplicative factor.

git-svn-id: http://130.66.124.6/svn/eonav/oad_and_monitoring_modules@429 9358c2ab-9c44-400d-ad1c-7153579a8fc0