Fixed a problem with DataSource's copy constructor
When copying a DataSource, if nothing is done, all the modules will
maintain a pointer to the copied DataSource: we need to update the
pointer maintained by each module to point to the new DataSource. This
bug was discovered when creating DataSourceDerivatives class.
We also need to make sure that the copy is independant from the
original.

git-svn-id: http://130.66.124.6/svn/dev_pilotage/simulator_sdk/trunk/data_source@337 5574b880-4a8e-4d59-90b7-38b80912c59c