Stubbed 'Cumulate' class
This class will be used to calculate the sum (integral) of a variable
over time. Unlike the other classes in that module (Simpson, Quadpack,
etc.), it allows adding points to the function being integrated. The
function is not defined analytically but by (x,y) pairs. The integration
scheme is trapezoidal integration.

git-svn-id: http://130.66.124.6/svn/eonav/oad_and_monitoring_modules@936 9358c2ab-9c44-400d-ad1c-7153579a8fc0