Added Herb Sutter's generic pimpl
This needs C++11 (at least gcc 4.3).

Quoted from http://herbsutter.com/gotw/_101/:
Why is this an improvement over the hand-rolled Pimpl Idiom?

First, the code is simpler because it eliminates some pieces of
boilerplate: In the hand-rolled version, you also have to declare the
constructor and write its body in the implementation file and explicitly
allocate the impl object. You also have to remember to declare the
destructor and write its body in the implementation file, for obscure
language reasons explained in #100.
Second, the code is more robust: In the hand-rolled version, if you
forget to write the out-of-line destructor, the Pimpld class will
compile in isolation and appear to be in a check-in-able state, but will
fail to compile when used by a caller that tries to destroy an object
and encounters a helpful cannot generate destructor because impl is,
uh, you know, incomplete error that leaves the author of the calling
code scratching his head as he walks over to your office to ream you out
for checking in something broken.

git-svn-id: http://130.66.124.6/svn/eonav/oad_and_monitoring_modules@938 9358c2ab-9c44-400d-ad1c-7153579a8fc0