Refactored loxodrome_on_ellipsoid
This was done to prepare the implementation of the direct loxodrome problem
resolution: the Newton-Raphson method will need to evaluate the derivative of
the series, hence we must separate the computation of the coefficients from
their use in calculating either the series or its derivative.
This comes at a price, as we lose about 10% of computation time in this process.

git-svn-id: http://130.66.124.6/svn/eonav/oad_and_monitoring_modules@1244 9358c2ab-9c44-400d-ad1c-7153579a8fc0