Changed the way a Leg is constructed.
The length of the Leg cannot be computed by the constructor, as it depends on
the type of Leg & we cannot call a virtual method from a constructor.
Therefore, we used a static builder paradigm: the only available constructor
for ShortestPathLeg takes the length as argument. Therefore, to make a
ShortestPathLeg instance, we must call the static 'build' method, which
computes the leg length.

git-svn-id: http://130.66.124.6/svn/eonav/oad_and_monitoring_modules@1262 9358c2ab-9c44-400d-ad1c-7153579a8fc0