Fix: added lazy initialization to DecodeUnit
This was done because C++ doesn't allow ordering the initialization
of static variables in separate compilation units: therefore, when
defining a static variable as the result of a DecodeUnit, the
program crashes if DecodeUnit is initialized after. This fix
prevents this proble & DecodeUnit can now safely be used with static
variables.

git-svn-id: http://130.66.124.6/svn/maroff/Product/Cpp/Tools/decode_unit@1718 300066c3-8af3-463d-8cc0-c0233cc64ac1