Fixed DecodeUnit's initialization
The problem with the lazy evaluation as it was implemented
was that it created memory leaks as we could overwrite
a static variable before it was initialize (typically,
when using the result of DecodeUnit to initialize another
static variable).
The new implementation uses a static variable inside
static methods & separates the small number of known units
(at compile-time) from those calculated during the
bootstrap (using those known units).

Now DecodeUnits has no memory leaks & can be used to
initialize static variables (as was done in the code
of the IRT's simulator to prevent duplicate parsing
of YAML test data).

git-svn-id: http://130.66.124.6/svn/maroff/Product/Cpp/Tools/decode_unit@1721 300066c3-8af3-463d-8cc0-c0233cc64ac1