Created configurable headers for each module
These headers will be used to include all the headers of a module.
The aim is to make the modules as much "file-agnostic" as possible: the clients
will depend on the actual class & function names, but not on the filenames of
the headers (because they'll be masked by the module's signle header file).
Much like boost, really.